Add path aliases Renamed .d.ts extensions to .ts Updated import paths
