+++ src/app/components/layout/header/header.component.ts
@@ -1,7 +1,7 @@
 import { Component, DoCheck } from '@angular/core';
-import { AuthService } from 'src/app/services/auth/auth.service';
 import { RouterLink, RouterLinkActive } from '@angular/router';
 import { NgIf } from '@angular/common';
+import { AuthService } from '@services/auth/auth.service';
 
 @Component({
   selector: 'app-header',
