+++ src/app/components/login/login.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { LoginComponent } from './login.component';
-import { AuthService } from 'src/app/services/auth/auth.service';
 import { ReactiveFormsModule } from '@angular/forms';
+import { AuthService } from '@services/auth/auth.service';
 
 describe('LoginComponent', () => {
   let component: LoginComponent;
