+++ src/app/services/app-insights/app-insights.service.spec.ts
@@ -1,5 +1,5 @@
 import { ApplicationInsights } from '@microsoft/applicationinsights-web';
-import { AppConfigService } from '../app-config/app-config.service';
+import { AppConfigService } from '@services/app-config/app-config.service';
 import { AppInsightsService } from './app-insights.service';
 
 describe('AppInsightsService', () => {
