+++ src/app/types/hearing.ts
@@ -1,7 +1,7 @@
 export interface HearingData {
   id: number;
   date: string;
-  judges: string[] = [];
+  judges: string[];
   courtroom: string;
   transcript_count: number;
 }
