+++ src/main.ts
@@ -1,13 +1,13 @@
 import { AppComponent } from './app/components/app.component';
 import { withInterceptorsFromDi, provideHttpClient, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { bootstrapApplication } from '@angular/platform-browser';
-import { ErrorHandlerService } from './app/services/error/error-handler.service';
-import { ErrorInterceptor } from './app/interceptor/error/error.interceptor';
-import { AppConfigService } from './app/services/app-config/app-config.service';
 import { APP_INITIALIZER, ErrorHandler } from '@angular/core';
 import { provideRouter } from '@angular/router';
 import { APP_ROUTES } from './app/app.routes';
-import { AppInsightsService } from './app/services/app-insights/app-insights.service';
+import { ErrorInterceptor } from '@interceptors/error/error.interceptor';
+import { AppConfigService } from '@services/app-config/app-config.service';
+import { AppInsightsService } from '@services/app-insights/app-insights.service';
+import { ErrorHandlerService } from '@services/error/error-handler.service';
 
 export function initAppFn(envService: AppConfigService) {
   return () => envService.loadAppConfig();
